+++ src/importscraper.h
@@ -35,6 +35,7 @@ class ImportScraper : public AbstractScraper {
 
 public:
     ImportScraper(Settings *config, QSharedPointer<NetManager> manager);
+
     void runPasses(QList<GameEntry> &gameEntries, const QFileInfo &info,
                    QString &, QString &) override;
     void getGameData(GameEntry &game) override;