+++ src/openretro.cpp
@@ -35,7 +35,7 @@ OpenRetro::OpenRetro(Settings *config, QSharedPointer<NetManager> manager)
     : AbstractScraper(config, manager, MatchType::MATCH_MANY) {
     baseUrl = "https://openretro.org";
 
-    searchUrlPre = "https://openretro.org";
+    searchUrlPre = baseUrl;
     searchUrlPost = "&unpublished=1";
 
     searchResultPre = "<div style='margin-bottom: 4px;'>";